requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.3.1
cursor==1.2.0
decorator==4.3.0
dogpile.cache==0.6.7
enum34==1.1.6
halo==0.0.21
idna==2.7
iso8601==0.1.12
jmespath==0.9.3
jsonpatch==1.23
jsonpointer==2.0
keystoneauth1==3.11.1
log-symbols==0.0.11
munch==2.3.2
netifaces==0.10.7
openstacksdk==0.19.0
os-service-types==1.3.0
pbr==5.1.0
pycparser==2.19
PyYAML==3.13
requests==2.20.0
requestsexceptions==1.4.0
six==1.11.0
spinners==0.0.23
stevedore==1.30.0
termcolor==1.1.0
urllib3==1.24
